move all node classes into the node module
